# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sabor & Gestión API
  description: API REST del sistema Sabor & Gestión para la gestión operativa del restaurante.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo local
...
paths:
...
  /api/categorias:
    get:
      tags:
        - Categorías
      summary: Listar categorías del menú
      description: Endpoint público para consultar todas las categorías registradas en el menú.
      operationId: obtenerCategorias
      responses:
        '200':
          description: Lista de categorías obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Categoria'
        '500':
          description: Error al obtener las categorías
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
    post:
      tags:
        - Categorías
      summary: Crear una nueva categoría
      description: Crea una categoría de menú. Requiere rol de Administrador. Valida que el nombre contenga solo letras y no esté duplicado.
      operationId: crearCategoria
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoriaRequest'
      responses:
        '201':
          description: Categoría creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Categoria'
        '400':
          description: Nombre inválido o duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '401':
          description: No autenticado o token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '403':
          description: Acceso denegado (requiere Administrador)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al crear la categoría
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
  /api/categorias/{id}:
    put:
      tags:
        - Categorías
      summary: Actualizar una categoría existente
      description: Actualiza el nombre de una categoría. Requiere rol de Administrador.
      operationId: actualizarCategoria
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador de la categoría
          schema:
            type: string
          example: '{{categoriaId}}'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoriaRequest'
      responses:
        '200':
          description: Categoría actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Categoria'
        '400':
          description: Nombre inválido o duplicado con otra categoría
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '401':
          description: No autenticado o token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '403':
          description: Acceso denegado (requiere Administrador)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '404':
          description: Categoría no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al actualizar categoría
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
    delete:
      tags:
        - Categorías
      summary: Eliminar una categoría
      description: Elimina una categoría por su identificador. Requiere rol de Administrador.
      operationId: eliminarCategoria
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador de la categoría
          schema:
            type: string
          example: '{{categoriaId}}'
      responses:
        '200':
          description: Categoría eliminada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeRespuestaResponse'
              example:
                mensaje: Categoría eliminada correctamente
        '401':
          description: No autenticado o token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '403':
          description: Acceso denegado (requiere Administrador)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '404':
          description: Categoría no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al eliminar categoría
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
...
  /api/mesas:
    get:
      tags:
        - Mesas
      summary: Listar todas las mesas
      description: Obtiene las mesas del restaurante con su ubicación poblada y estado mapeado (Disponible, Ocupada, Esperando pago). Admite filtrado por ubicación. Requiere autenticación.
      operationId: obtenerMesas
      security:
        - bearerAuth: []
      parameters:
        - name: location
          in: query
          required: false
          description: Filtrar por ID o nombre de ubicación
          schema:
            type: string
          example: '{{ubicacionId}}'
      responses:
        '200':
          description: Lista de mesas obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MesaDTO'
        '401':
          description: No autenticado o token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al obtener mesas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
    post:
      tags:
        - Mesas
      summary: Crear mesa o lote de mesas
      description: Crea una mesa o un lote de mesas. Requiere rol Administrador. Acepta nomenclatura bilingüe (name/numero, capacity/capacidad, location/ubicacion, type/tipo, status/estado).
      operationId: crearMesa
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/CrearMesaRequest'
                - type: array
                  items:
                    $ref: '#/components/schemas/CrearMesaRequest'
      responses:
        '201':
          description: Mesa o lote creado exitosamente
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/MesaDTO'
                  - type: array
                    items:
                      $ref: '#/components/schemas/MesaDTO'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '401':
          description: No autenticado o token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '403':
          description: Acceso denegado (requiere Administrador)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al crear la mesa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
  /api/mesas/{id}:
    get:
      tags:
        - Mesas
      summary: Obtener detalle de una mesa
      description: Obtiene los datos detallados de una mesa por su ID. Requiere autenticación.
      operationId: obtenerMesaPorId
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador de la mesa
          schema:
            type: string
          example: '{{mesaId}}'
      responses:
        '200':
          description: Detalle de la mesa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MesaDTO'
        '401':
          description: No autenticado o token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '404':
          description: Mesa no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al obtener la mesa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
    put:
      tags:
        - Mesas
      summary: Actualizar campos generales de una mesa
      description: Actualiza nombre, capacidad, tipo o ubicación de una mesa. Requiere rol de Administrador. Valida regla de formato de nombre de mesa.
      operationId: actualizarMesa
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador de la mesa
          schema:
            type: string
          example: '{{mesaId}}'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActualizarMesaRequest'
      responses:
        '200':
          description: Mesa actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MesaDTO'
        '400':
          description: Regla de formato de nombre fallida o nombre ya registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '401':
          description: No autenticado o token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '403':
          description: Acceso denegado (requiere Administrador)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '404':
          description: Mesa no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al actualizar mesa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
    delete:
      tags:
        - Mesas
      summary: Eliminar una mesa
      description: Elimina una mesa por ID. Requiere rol Administrador.
      operationId: eliminarMesa
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador de la mesa
          schema:
            type: string
          example: '{{mesaId}}'
      responses:
        '200':
          description: Mesa eliminada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EliminarMesaResponse'
        '401':
          description: No autenticado o token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '403':
          description: Acceso denegado (requiere Administrador)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '404':
          description: Mesa no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al eliminar mesa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
  /api/mesas/{id}/estado:
    patch:
      tags:
        - Mesas
      summary: Actualizar solo el estado de una mesa
      description: 'Actualiza el estado operativo de una mesa. Valores aceptados: "Libre", "Ocupada", "Cuenta Solicitada". Requiere autenticación.'
      operationId: actualizarEstadoMesa
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador de la mesa
          schema:
            type: string
          example: '{{mesaId}}'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActualizarEstadoMesaRequest'
      responses:
        '200':
          description: Estado de la mesa actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MesaDTO'
        '401':
          description: No autenticado o token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '404':
          description: Mesa no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al actualizar estado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
...
  /api/dashboard/resumen:
    get:
      tags:
        - Dashboard
      summary: Obtener resumen, KPIs y métricas del negocio
      description: Retorna los KPIs del día (ventasHoy, ordenesHoy, clientesEstimados, mesasActivas, ocupacionPorcentaje), platos más vendidos, categorías populares y órdenes recientes. Requiere rol Administrador.
      operationId: obtenerResumenDashboard
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Resumen consolidado del dashboard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumenDashboardDTO'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '403':
          description: Acceso denegado (requiere Administrador)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al generar resumen del Dashboard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Autenticación mediante token JWT en cabecera: Authorization: Bearer <token>'
  schemas:
    MensajeErrorResponse:
      type: object
      required:
        - mensaje
      properties:
        mensaje:
          type: string
    MensajeRespuestaResponse:
      type: object
      required:
        - mensaje
      properties:
        mensaje:
          type: string
          example: Operación realizada exitosamente
...
    Categoria:
      type: object
      properties:
        id:
          type: string
          example: '000000000000000000000010'
        _id:
          type: string
          example: '000000000000000000000010'
        nombre:
          type: string
          example: Platos Fuertes
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CategoriaRequest:
      type: object
      required:
        - nombre
      properties:
        nombre:
          type: string
          example: Bebidas
...
    MesaDTO:
      type: object
      properties:
        id:
          type: string
          example: '000000000000000000000030'
        _id:
          type: string
          example: '000000000000000000000030'
        name:
          type: string
          example: Mesa 1 Interior
        numero:
          type: string
          example: Mesa 1 Interior
        capacity:
          type: integer
          example: 4
        location:
          type: string
          example: Interior
        locationId:
          type: string
          nullable: true
          example: '000000000000000000000040'
        status:
          type: string
          example: Disponible
          description: Mapeo para visualización en frontend y eventos WebSocket (Disponible, Ocupada, Esperando pago)
        type:
          type: string
          enum:
            - normal
            - vip
          example: normal
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CrearMesaRequest:
      type: object
      properties:
        name:
          type: string
          example: Mesa 5 Interior
        numero:
          type: string
          example: Mesa 5 Interior
        capacity:
          type: integer
          example: 4
        capacidad:
          type: integer
          example: 4
        location:
          type: string
          example: Interior
        ubicacion:
          type: string
          example: Interior
        type:
          type: string
          enum:
            - normal
            - vip
          example: normal
        tipo:
          type: string
          enum:
            - normal
            - vip
          example: normal
        status:
          type: string
          example: Libre
          description: Alias opcional en inglés para el estado de la mesa
        estado:
          type: string
          enum:
            - Libre
            - Ocupada
            - Cuenta Solicitada
          example: Libre
    ActualizarMesaRequest:
      type: object
      properties:
        name:
          type: string
          example: Mesa 5 Terraza
        numero:
          type: string
          example: Mesa 5 Terraza
        capacity:
          type: integer
          example: 6
        capacidad:
          type: integer
          example: 6
        location:
          type: string
          example: Terraza
        ubicacion:
          type: string
          example: Terraza
        type:
          type: string
          enum:
            - normal
            - vip
          example: vip
        tipo:
          type: string
          enum:
            - normal
            - vip
          example: vip
        status:
          type: string
          example: Libre
          description: Alias opcional en inglés para el estado de la mesa
        estado:
          type: string
          enum:
            - Libre
            - Ocupada
            - Cuenta Solicitada
          example: Libre
    ActualizarEstadoMesaRequest:
      type: object
      required:
        - estado
      properties:
        estado:
          type: string
          enum:
            - Libre
            - Ocupada
            - Cuenta Solicitada
          example: Ocupada
          description: Estado operativo de la mesa en MongoDB
    EliminarMesaResponse:
      type: object
      properties:
        mensaje:
          type: string
          example: Mesa eliminada
        mesa:
          $ref: '#/components/schemas/MesaDTO'
...
    KPIsDashboard:
      type: object
      properties:
        ventasHoy:
          type: number
          example: 3450
        ordenesHoy:
          type: integer
          example: 28
        clientesEstimados:
          type: integer
          example: 75
        mesasActivas:
          type: integer
          example: 6
        ocupacionPorcentaje:
          type: number
          example: 60
    PlatoPopularDTO:
      type: object
      properties:
        nombre:
          type: string
          example: Pique Macho
        cantidad:
          type: integer
          example: 14
        imagen:
          type: string
          example: https://res.cloudinary.com/demo/image/upload/pique.jpg
    CategoriaPopularDTO:
      type: object
      properties:
        nombre:
          type: string
          example: Platos Fuertes
        pedidos:
          type: integer
          example: 20
        porcentaje:
          type: number
          example: 71.4
    OrdenRecienteDTO:
      type: object
      properties:
        id:
          type: string
          example: PED-0001
        mesa:
          type: string
          example: Mesa 3 Interior
        hora:
          type: string
          example: '14:25'
        estado:
          type: string
          example: En Cocina
        total:
          type: number
          example: 120
    ResumenDashboardDTO:
      type: object
      properties:
        kpis:
          $ref: '#/components/schemas/KPIsDashboard'
        platosMasVendidos:
          type: array
          items:
            $ref: '#/components/schemas/PlatoPopularDTO'
        categoriasPopulares:
          type: array
          items:
            $ref: '#/components/schemas/CategoriaPopularDTO'
        ordenesRecientes:
          type: array
          items:
            $ref: '#/components/schemas/OrdenRecienteDTO'